% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{chainedLaspeyres}
\alias{chainedLaspeyres}
\title{Function to generate chained Laspeyres index}
\usage{
chainedLaspeyres(weights, exchangerates)
}
\arguments{
\item{weights}{tibble with yearly weights where each row is a year and each column is the weight of a given country. The column names of the country weights must be equal to those of the exchangerates column names. In addition, a coulmn with years must be included, named year}

\item{exchangerates}{tibble with monthly exchange rates where each row is a month and each column is the exchange rate of a currency. The column names of the currencies must be equal to those of the weights column names. In addition, a coulmn with months must be included, named date_start.}
}
\value{
tibble with dates and index
}
\description{
EXPECT MAJOR CHANGES
}
\examples{

#Soon
}
