% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggnewsPlot}
\alias{ggnewsPlot}
\title{Replicates DN's RevPAR figure in ggplot2}
\usage{
ggnewsPlot(df, outputReady = FALSE, customFont = FALSE)
}
\arguments{
\item{df}{input data.frame containg place, revpar and change in decimals.}

\item{outputReady}{Logical. If TRUE, uses device size and is better suited for ouputs. Default is FALSE - fixed aspect ration.}

\item{customFont}{Logical. If TRUE, use the font family Montserrat (if available). Default is 'sans'}
}
\value{
ggplot object
}
\description{
The function generates a figure often used by Dagens Næringsliv (DN), a Norwegian newspaper specializing in business news,
to present data on RevPAR and changes in RevPAR for diferent regions/cities.
}
\details{
This function is early developement. Expect revisions.
}
\examples{

exampleFrame <- data.frame(Sted = c("Tromsø", "Stanvager", "Bodø", "Trondheim", "Bergen", "Lillehammer", "Gardermoen", "Haugesund", "Kristiansund", "Oslo", "Kristiansand"),
                           RevPAR = c(925, 432, 605, 526, 391, 463, 601, 367, 370, 613, 335),
                           Endring = c(0.153, 0.137, 0.036, 0.014, 0.008, -0.005, -0.029, -0.054, -0.074, -0.082, -0.179))

ggnewsPlot(exampleFrame)
}
