% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{giniCor}
\alias{giniCor}
\title{Estimate the Gini correlation}
\usage{
giniCor(y, x)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}
}
\value{
value betweel -1 and 1
}
\description{
Estimate the Gini correlation
}
\examples{
x1 <- c(0,0,1,1,1,1.2,1.2,1.2,1,1,1,0)
x2 <- c(0.4,0.4,0.4,0,0,0,0,0,0,0,0,0.4)
y <- x1+x2

giniCor(y, x1)
}
