% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{giniMonths}
\alias{giniMonths}
\title{Calculate months equivalent degree of tourism seasonality}
\usage{
giniMonths(gini)
}
\arguments{
\item{gini}{numeric value between 0 and (n-1)/n}
}
\value{
numeric value between 0 and 12
}
\description{
\deqn{12 \times (1-G_y)}
See Sitouras (2004).
}
\examples{

x <- c(1:12)
giniMonths(gini(x))
}
