% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laspeyresCurrencies}
\alias{laspeyresCurrencies}
\title{Exchange rates}
\format{A tibble with 167 rows and 22 columns
\describe{
  \item{date_start}{Month of observation}
  \item{year}{Year of observation}
  \item{...}{Currency}
}}
\source{
\url{https://www.norges-bank.no}
}
\usage{
laspeyresCurrencies
}
\description{
Exchange rates for 20 currencies against the NOK, 2007-2020
}
\keyword{datasets}
