% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laspeyresWeights}
\alias{laspeyresWeights}
\title{Yearly guest night weights}
\format{A tibble with 15 rows and 21 columns:
\describe{
  \item{year}{Year of observation}
  \item{...}{country of origin}
}}
\source{
\url{https://www.ssb.no}
}
\usage{
laspeyresWeights
}
\description{
Yearly guest nights weights by country of origin to Norway, 2006-2020
}
\keyword{datasets}
