% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norwayGuestNights2018}
\alias{norwayGuestNights2018}
\title{Guest nights Norway 2018}
\format{A data frame with 12 rows and 20 variables:
\describe{
  \item{month}{Month of observation}
  \item{LOCATIONS}{Counties in Norway}
  \item{...}{...}
}}
\source{
\url{https://www.ssb.no/}
}
\usage{
norwayGuestNights2018
}
\description{
Guest nights in Norway for the year 2018 by county
}
\keyword{datasets}
