% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{RME}
\alias{RME}
\title{Estimate the RME}
\usage{
RME(y, x)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}
}
\value{
numeric
}
\description{
RME measures the marginal effect each G_k has on G_y where sum k = 1 to n (G_k) = G_y
}
\examples{
#SOON
}
