% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{themeEmpty}
\alias{themeEmpty}
\title{Empty theme ggplot}
\usage{
themeEmpty()
}
\value{
ggtheme
}
\description{
Removes all text, gridlines, background color etc.
}
\examples{
 ggplot2::ggplot(data.frame(x=rep(c(1, 2, 3, 4, 5), 2), y=1:10, d=rep(c("a", "b"), 5)), ggplot2::aes(x=x, y=y, fill=d))+
  ggplot2::geom_bar(stat="identity", position="dodge")+
  themehemeEmpty()
}
