% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tromso}
\alias{tromso}
\title{Cruise passengers to Tromsø harbour}
\format{A data frame with 49 rows and 2 variables:
\describe{
  \item{date_start}{Month of observation}
  \item{Passasjerer}{Number of passengers}
}}
\source{
\url{https://www.kystverket.no/safeseanet}
}
\usage{
tromso
}
\description{
A time series of cruise passengers to Tromsø harbour
May 2015 - May 2019.
}
\keyword{datasets}
