% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Estimate the Gini Coefficient}
\usage{
gini(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
value between 0 and (n-1)/n
}
\description{
Gini coefficient estimated using the covariance approach
See Lerman and Yitzaki (1989) and A. Fernandez-Morales et al. (2016)
}
\examples{
#SOON
}
