% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{gini}
\alias{gini}
\title{Estimate the Gini Coefficient}
\usage{
gini(x, suppressWarning = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{suppressWarning}{logical. Set \code{TRUE} to hide \code{length(x) != 12} warning}
}
\value{
value between 0 and (n-1)/n
}
\description{
Gini coefficient estimated using the covariance approach, i.e.
\deqn{G_y = \frac{2 cov (y,F(y))}{\mu_y}}.
See Lerman and Yitzaki (1989) and A. Fernandez-Morales et al. (2016)
}
\details{
Currently returns 0 if x is only zeroes.
}
\examples{

x <- c(1:12)
gini(x)
}
