% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{giniCor}
\alias{giniCor}
\title{Estimate the Gini correlation}
\usage{
giniCor(x1, x2)
}
\arguments{
\item{x1}{numeric vector}

\item{x2}{numeric vector}
}
\value{
value betweel -1 and 1
}
\description{
Estimate the Gini correlation
}
\examples{
#SOON
}
