% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini.R
\name{giniRME}
\alias{giniRME}
\title{Estimate the RME}
\usage{
giniRME(y, x)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}
}
\value{
numeric
}
\description{
RME measures the marginal effect each G_k has on G_y where sum k = 1 to n (G_k) = G_y
}
\examples{
x1 <- c(0,0,1,1,1,1.2,1.2,1.2,1,1,1,0)
x2 <- c(0.4,0.4,0.4,0,0,0,0,0,0,0,0,0.4)
y <- x1+x2

giniRME(y, x1)
}
